fullCalendar's today should respect faked now.
