+++ course/templates/course/calendar.html
@@ -175,6 +175,7 @@
           },
           contentHeight: "auto",
           defaultDate: '{{ default_date }}',
+          now: '{{ now_date }}',
           defaultView: $(window).width() < 765 ? 'listMonth': 'month',
           timezone: "local",
           nowIndicator: true,
