+++ relate/templates/modal-form.html
@@ -4,7 +4,7 @@
 
 <div class="modal fade" id="{{ form.modal_id}}" tabindex="-1" role="dialog"
      aria-labelledby="{{ form.modal_id}}_label" aria-hidden="true">
-  <div class="modal-dialog modal-lg modal-sm"">
+  <div class="modal-dialog modal-lg modal-sm">
     <div class="modal-content">
       <div class="modal-header">
         <button type="button" class="close" data-dismiss="modal">
@@ -17,3 +17,11 @@
     </div>
   </div>
 </div>
+
+{% if extra_js %}
+  <script>
+    (function() {
+      {{ extra_js |safe }}
+    })();
+  </script>
+{% endif %}
